Adding cache variable Optika_DO_GUI_UNIT_TESTS (bug 5563)
This updated CMake code keeps the same configure-time logic of whether to
enable or not the Optika GUI tests based on various logic.  However, this new
option Optika_DO_GUI_UNIT_TESTS allows the user to disable the GUI unit tests
to overrride this choice.

See Trilinos Bugzilla bug 5563 for more details on this.